Add timestamped submodule
